+++ src/cpp/core/include/core/LogOptions.hpp
@@ -24,6 +24,8 @@
 #include <core/Log.hpp>
 #include <core/Thread.hpp>
 
+#include <core/system/Types.hpp>
+
 #include <shared_core/FilePath.hpp>
 #include <shared_core/FileLogDestination.hpp>
 
@@ -51,12 +53,16 @@ public:
    LogOptions(const std::string& executableName,
               LogLevel defaultLogLevel,
               LoggerType defaultLoggerType,
+              LogMessageFormatType defaultMessageFormatType,
               const LoggerOptions& defaultLoggerOptions);
 
    virtual ~LogOptions()
    { }
 
    core::Error read();
+
+   // gets the current log message format type
+   LogMessageFormatType logMessageFormatType(const std::string& loggerName = std::string()) const;
 
    // gets the current log level
    LogLevel logLevel(const std::string& loggerName = std::string()) const;
@@ -72,6 +78,8 @@ public:
 
    std::vector<std::string> loggerOverrides() const;
 
+   static FilePath defaultLogDirectory();
+
 private:
    void initProfile();
 
@@ -83,12 +91,15 @@ private:
 
    std::string defaultLogLevel_;
    std::string defaultLoggerType_;
+   std::string defaultMessageFormatType_;
    LoggerOptions defaultLoggerOptions_;
 
    LogLevel lowestLogLevel_;
 
    ConfigProfile profile_;
 };
+
+void forwardLogOptionsEnvVars(core::system::Options* pEnvironment);
 
 } // namespace log
 } // namespace core