+++ src/cpp/session/postback/PostbackMain.cpp
@@ -54,7 +54,7 @@ int main(int argc, char * const argv[])
    {
       // initialize log
       core::log::setProgramId("rpostback");
-      core::system::initializeSystemLog("rpostback", core::log::LogLevel::WARN);
+      core::system::initializeLog("rpostback", core::log::LogLevel::WARN);
 
       // ignore SIGPIPE
       Error error = core::system::ignoreSignal(core::system::SigPipe);